requirements.txt
nltk==3.7
Pillow==10.0.0
torch==2.0.1
torchtext==0.15.2
torchvision==0.10.0
